Fixed a bug that allowed gun reload after endGame
Fixed a bug that allowed the gun to reload even after the game was over
and the end game menu was displayed. Added ground hog pictures to the
resources and tried to display them on the bonus level, but a bug
prevents them from being displayed. Temporarily reverted to black and
white background display.